fix types for precision check

+++ benchmark-runner/universalrunner/control_types.hpp
@@ -6,7 +6,7 @@
 #ifndef FLOAT
 #include <cnl/fixed_point.h>
 #else
-#include <half.hpp>
+#include "half.hpp"
 #endif
 
 typedef double __plant_precisiont;
@@ -16,13 +16,13 @@ typedef double __plant_precisiont;
 #ifndef FLOAT
 
 #if CONTROL_WIDTH == 8u
-typedef uint8_t __controller_base_typet;
+typedef int8_t __controller_base_typet;
 #elif CONTROL_WIDTH == 16u
-typedef uint16_t __controller_base_typet;
+typedef int16_t __controller_base_typet;
 #elif CONTROL_WIDTH == 32u
-typedef uint32_t __controller_base_typet;
+typedef int32_t __controller_base_typet;
 #elif CONTROL_WIDTH == 64u
-typedef uint32_t __controller_base_typet;
+typedef int64_t __controller_base_typet;
 #else
 #error "Unsupported controller width."
 #endif
